cpu_input_boost: Fight back against the user
There was a user_policy vs. policy condition occurring for online CPUs where the minfreq of said online CPUs were permanently stuck at the boost frequency because the end user attempted to modify the minfreq via scaling_min_freq (user_policy). User_policy trumped kernel policy after the boost was over, resulting in the stuck minfreq.

To fix this, user_policy is used to boost online CPUs and the cpufreq notifier (kernel policy) is used to boost offline CPUs that are plugged back in during the duration of the boost.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>